Fix `npm install` by upgrading `nan` module
I was unable to `npm install` on my OSX Yosemite laptop due to a failure
compiling the `utf-8-validate` package. This depended on an old version
of `nan`, and upgrading that package seems to have alleviated the build
problem.

I also added a `.node-version` file which pins the version of NodeJS on
Travis and for developers who use nodenv/nvm.