Refactor committeePage
Rename the main state to be about contributions. Use the commiteePage.html
template in the abstract parent state as a ui-view container.
appMain.committee.contributions should render the committee-listing component
which should be able to be re-used for expenditure data as well (despite the
data still being called "contributions".

Some work probably still needs to be altered for committee-listing to be truly
re-usable but the one doing that work would know best.